Improves query 2
Query 2 now shows each user alongside with its profile and available
services. Its not ideal because, for each service available,
we repeat the users name and its profile. Dont know if this is a
issue with how we model our DB or just something that we can patch when
dispalying the query

Signed-off-by: caio <caiotdandrade@gmail.com>